rename files to fix weird sass import on tsx name collision, upgrade react-scripts
