+++ package.json
@@ -40,7 +40,7 @@
     "react-redux": "^7.1.1",
     "react-router": "^5.1.0",
     "react-router-dom": "^5.1.0",
-    "react-scripts": "3.2.0",
+    "react-scripts": "3.3.0",
     "react-toastify": "^5.4.1",
     "redux-oidc": "^3.1.5",
     "redux-persist": "^6.0.0",
