+++ src/domain/app/layout/PageLayout.tsx
@@ -2,7 +2,7 @@ import React, { FunctionComponent } from 'react';
 import { HelmetProvider } from 'react-helmet-async';
 
 import Header from '../header/Header';
-import styles from './layout.module.scss';
+import styles from './pageLayout.module.scss';
 import Footer from '../footer/Footer';
 const PageLayout: FunctionComponent = ({ children }) => {
   return (
