+++ src/domain/app/layout/__tests__/Layout.test.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
 
-import Layout from '../Layout';
+import PageLayout from '../PageLayout';
 
 it('renders snapshot correctly', () => {
   const layout = shallow(
-    <Layout>
+    <PageLayout>
       <></>
-    </Layout>
+    </PageLayout>
   );
   expect(toJson(layout)).toMatchSnapshot();
 });
